TEST: add global CFLAGS for graphite loop optimizer
this is a native loop optimizer in the GCC 4.6+ tool chain.
all i've done is enable it to be used.  Boardconfig.mk should contain
the edit to enable this per device.  I set it up this way until tested
on all builds because we ran into drawing issues the last time we
hotrodded webkit with jpeg-turbo under strict aliasing which we have
set globally.

ps2: make sure we hit arm thumbs too!

Change-Id: I11d7d7cde87582216f717de947945fad16ed2cf7